Concentrate validated OBD target evidence
